feat!: Add Integer type
Adds a new `Integer` type to toml_edit. This represents a TOML integer
and its possible notations:

- Decimal: Base 10, can be any valid i64
- Hexadecimal: Base 16, must be nonnegative.
    - Hexadecimal integers can be uppercase or lowercase.
      This only applies to user-created Integers; those parsed from TOML
      strings always have the `Unspecified` case.
      This covers most use cases while avoidng the complexity of
      supporting customizable alphabet casing.
- Octal: Base 8, must be nonnegative.
- Binary: Base 2, must be nonnegative.

The `Integer` type is modeled after the `Datetime` type: it provides no
mutating methods, and Value does not expose a mutable reference for
Integer. This should be OK since Value does not expose other primitive
values (f64, Datetime, etc.) as mutable references either.

Note: I kept the `impl From<i64> for Value` for convenience.

+++ crates/toml_edit/src/encode.rs
@@ -13,7 +13,7 @@ use crate::value::{
     DEFAULT_LEADING_VALUE_DECOR, DEFAULT_TRAILING_VALUE_DECOR, DEFAULT_VALUE_DECOR,
 };
 use crate::DocumentMut;
-use crate::{Array, InlineTable, Item, Table, Value};
+use crate::{Array, InlineTable, Integer, Item, Table, Value};
 
 pub(crate) fn encode_key(this: &Key, buf: &mut dyn Write, input: Option<&str>) -> Result {
     if let Some(input) = input {
@@ -513,7 +513,7 @@ fn infer_all_style(value: &str) -> (StringStyle, bool) {
     (ty, true)
 }
 
-impl ValueRepr for i64 {
+impl ValueRepr for Integer {
     fn to_repr(&self) -> Repr {
         Repr::new_unchecked(self.to_string())
     }
